refactor: export ignition errors on their own entry point
The two Ignition errors are now defined at the root under `./errors.ts`.
The supporting utility functions remain under the
`./internal/utils/paramAssertions.ts` file.

The file containing the Ignition errors is exported as a valid entry
point. A third party can access them by:

```js
import { IgnitionError } from "@ignored/ignition-core/errors"
```

Part of #171.
