+++ packages/core/package.json
@@ -22,12 +22,16 @@
   "types": "./dist/src/index.d.ts",
   "exports": {
     ".": "./dist/src/index.js",
-    "./helpers": "./dist/src/helpers.js"
+    "./helpers": "./dist/src/helpers.js",
+    "./errors": "./dist/src/errors.js"
   },
   "typesVersions": {
     "*": {
       "helpers": [
         "./dist/src/helpers"
+      ],
+      "errors": [
+        "./dist/src/errors"
       ]
     }
   },
