+++ packages/core/src/internal/execution/dispatch/utils.ts
@@ -3,7 +3,7 @@ import type {
   ExecutionResultsAccumulator,
 } from "../../types/executionGraph";
 
-import { IgnitionError } from "../../utils/errors";
+import { IgnitionError } from "../../../errors";
 import { isDependable, isEventParam, isProxy } from "../../utils/guards";
 
 export function toAddress(v: any) {
