+++ packages/core/src/internal/process/generateDeploymentGraphFrom.ts
@@ -5,8 +5,8 @@ import type {
 } from "../types/deploymentGraph";
 
 import { DeploymentBuilder } from "../../dsl/DeploymentBuilder";
+import { IgnitionError } from "../../errors";
 import { Module, ModuleDict } from "../../types/module";
-import { IgnitionError } from "../utils/errors";
 import { assertModuleReturnTypes } from "../utils/guards";
 
 export function generateDeploymentGraphFrom<T extends ModuleDict>(
