+++ packages/core/src/internal/validation/validateDeploymentGraph.ts
@@ -1,9 +1,9 @@
+import { IgnitionError } from "../../errors";
 import { getSortedVertexIdsFrom } from "../graph/utils";
 import { visit } from "../graph/visit";
 import { CallPoints, IDeploymentGraph } from "../types/deploymentGraph";
 import { Services } from "../types/services";
 import { ValidationVisitResult } from "../types/validation";
-import { IgnitionError } from "../utils/errors";
 
 import { validationDispatch } from "./dispatch/validationDispatch";
