+++ packages/core/test/helpers.ts
@@ -1,5 +1,6 @@
 import { ethers } from "ethers";
 
+import { IgnitionError } from "../src/errors";
 import {
   Services,
   TransactionOptions,
@@ -10,7 +11,6 @@ import {
   IArtifactsService,
   IAccountsService,
 } from "../src/internal/types/services";
-import { IgnitionError } from "../src/internal/utils/errors";
 import { Artifact } from "../src/types/hardhat";
 import { HasParamResult } from "../src/types/providers";
