+++ packages/core/test/validation.ts
@@ -6,10 +6,10 @@ import { ethers } from "ethers";
 import sinon from "sinon";
 
 import { buildModule } from "../src/dsl/buildModule";
+import { IgnitionValidationError } from "../src/errors";
 import { generateDeploymentGraphFrom } from "../src/internal/process/generateDeploymentGraphFrom";
 import { Services } from "../src/internal/types/services";
 import { ValidationVisitResult } from "../src/internal/types/validation";
-import { IgnitionValidationError } from "../src/internal/utils/errors";
 import { validateDeploymentGraph } from "../src/internal/validation/validateDeploymentGraph";
 import { ArtifactContract } from "../src/types/future";
 import { Artifact } from "../src/types/hardhat";
