+++ packages/hardhat-plugin/src/load-module.ts
@@ -1,5 +1,5 @@
 import { Module, ModuleDict } from "@ignored/ignition-core";
-import { IgnitionError } from "@ignored/ignition-core/helpers";
+import { IgnitionError } from "@ignored/ignition-core/errors";
 import setupDebug from "debug";
 import fsExtra from "fs-extra";
 import path from "path";
