+++ packages/hardhat-plugin/src/ui/components/execution/BatchExecution.tsx
@@ -1,5 +1,5 @@
 import { DeployState } from "@ignored/ignition-core";
-import { IgnitionError } from "@ignored/ignition-core/helpers";
+import { IgnitionError } from "@ignored/ignition-core/errors";
 import { Box, Text } from "ink";
 import Spinner from "ink-spinner";
