+++ packages/hardhat-plugin/src/ui/renderToCli.tsx
@@ -3,7 +3,7 @@ import {
   ModuleParams,
   UpdateUiAction,
 } from "@ignored/ignition-core";
-import { IgnitionError } from "@ignored/ignition-core/helpers";
+import { IgnitionError } from "@ignored/ignition-core/errors";
 import { render } from "ink";
 
 import { IgnitionUi } from "./components";
